fix vid63

+++ app/src/main/java/com/tofukma/shippingapp/common/Common.kt
@@ -132,8 +132,11 @@ object Common {
     fun getBearing(begin: LatLng, end: LatLng): Float {
         val lat = Math.abs(begin.latitude - end.longitude)
         val lng = Math.abs(begin.longitude - end.longitude)
+
         if(begin.latitude < end.latitude && begin.longitude < end.longitude)
-            return Math.toDegrees(Math.atan(lng/lat)).toFloat()
+            return Math.toDegrees(Math.atan(lng/lat))
+                .toFloat()
+
         else  if(begin.latitude >=  end.latitude && begin.longitude < end.longitude)
             return (90 - Math.toDegrees(Math.atan(lng/lat))+90).toFloat()
 
@@ -142,7 +145,7 @@ object Common {
 
         else if(begin.latitude < end.latitude && begin.longitude >= end.longitude)
             return (90 - Math.toDegrees(Math.atan(lng/lat))+270).toFloat()
-        return -1.0f
+        return (-1).toFloat()
     }
 
     fun decodePoly(encoded: String): List<LatLng> {
